Make new base class for xdsl-opt
